requirements.txt
Flask
transformers
sentence-transformers
torch
pdfminer.six
scikit-learn
numpy